ENH: Removing redundant data attributes from ParticleArray
The attributes removed are:

(a)temporary_arrays : Never knew the need for this

(b) property 'local': Local and  remote particles will be handled with the
    tag array so we don't need this.

(c) property 'group' : Don't know the need for this.

Also removed functions to set and remove 'flag' values. Again, I foresee
all such operations to work with the 'tag' array so we remove this for now.