ENH: Consistently setting num_particles and num_remote in NNPSParticleGeometric
This was being done from the earlier version of the code with reference to
the coordinate arrays. While this is not wrong, it makes reading the code a
little difficult. Essentially, at the end of lb_exchange_data and
remote_exchange_data, we know the num local and num remote particles
respecctively. This is where we should be setting it.