New way of writing equations.
This makes it much easier to write equations, eliminates need to specify
variables and temporaries, the code generation is much cleaner. The
Group class provides a lot of convenient methods.  The new equations
support pre-defined quantities.  Some simple code analysis is done to
simplify tasks.  A reasonable amount of testing is done for this and the
examples seem to run as well.

--HG--
branch : new_equation