ENH: Add Solver, Application and mayavi viewer from old PySPH.
This checkin adds the `Solver` and `Application` classes along with the
necessary support code.  It also adds the mayavi viewer and updates the
viewer to work with a recent mayavi version.  The dam break and
elliptical drop examples now run using the application.  Parallel
support still remains to be added.