ENH: Removing the redundant NNPS update at the end of the integration
Particles need only be binned when there is need to compute interactions from
neighbors. In serial, at the end of the integration step, we need to carry
forward the accelerations and use it to update to the half time step in the
next iteration.

In parallel, particles may be re-distributed after integration and again, we
should only require the previous time step's accelerations.