BUG: Fixing the parallel manager for multiple arrays.
We need to send pressure and sound speeds in the current form of the dam break
problem. This is beause the equation of state is called in the order of the
destinations. This means that a dummy value for source arrays is possible since
they have not been updated yet.

We can avoid sending these arrays if we have a more intelligent grouping of the
equations.