ENH: Adding kernels and updating application to properly setup NNPS
Kernels other than the cubic spline have different radius scale factors than the
default 2 which was being used to set up the NNPS object. A kernel radius
attribute is added to the kernels which is used by the Application in setting
up the NNPS object with the appropriate cell sizes.