BUG: Don't declare unassigned symbols.
Earlier, when wrapping a function, any undeclared symbols were being
declared as doubles, this would backfire when one was using a constant
like pi or M_1_PI or a builtin/math symbol and these would be set to
zero.  We used to filter out those but a cleaner solution is to only
declare any symbols that are assigned to in the code block.  This is now
implemented and tested.