ENH/BUG: Add initialize and fix post_loop.
The post_loop code was added at the end of each source which is
incorrect.  We now add it at the end after all sources for a particular
destination are done.  This allows us to fix a bug with the dam break
where we had to halve the acceleration due to gravity since there were
two sources and the post_loop was called twice.  Also added the
initialization function at the start of the destination loop.