ENH: Support for adaptive time stepping.
This adds basic support for adaptive time stepping.  Not all equations
or integrators support this.  It currently works for the dam-break but
there should be a way to turn adaptive timesteps on and off.  The
`DT_ADAPT` variable can be used to set the adaptive time stepping
factors inside any equation.  `DT_ADAPT` has three components `dt_cfl`,
`dt_force` and `dt_viscous`.