ENH: Optimizing the linked list NNPS for neighbor searching.
By constrution, the cell indices are always positive. This is used to
avoid spurious cells being considered for neighbor search. For
example, in the earlier version, due to the flattening routine, the
cell (0, 0, 0) would result in the spurious cell (-1, -1, 1) (which
doesn't exist) having a valid flattened index. This is now avoided.