BUG: Fixing particle iteration with LinkedList NNPS
For the linked list NNPS, the cells are constructed to have a positive
index. While searching for a particle's nearest neighbors, we need to
find the position of the particle relative to the minimum in the local
domain. This was not being done and as a result, the particle
iteration was correct only for the case when the physical coordinates
were >= 0.

The test has also been updated to handle this case.