API: Support for symmetric interactions.
This commit breaks the equation API. All the equations have been changed
so they can no longer accept multiple sources but only one source and
one destination.  One can additionally pass a symmetric boolean kwarg.
None of the equations themselves support symmetric interactions but the
framewokr supports this.  All the examples have been updated.  As the
equations start to support symmetric evaluation the examples can be
changed.  The one limitation is that all interactions between a src/dest
pair should be all symmetric or asymmetric if not, an error will be
raised.