ENH: Updating the adaptive time step option setting for PySPH
The option for adaptive time stepping and cfl number should be
available through the solver constructor. These options might then be
further overridden with command line options.

The CFL number was at a default value of 0.5 (set through the
integrator). This has been changed to be accessible from the solver
constructor and command line