BUG: Fixing the parallel initialization for multiple arrays
For parallel runs, the root creates all the particles, which is then
distributed using Zoltan. To correctly receive data on the other
processors, we needed an empty particle setup, using the particle info
created on the root. We were using a standard dict for this and that
resulted in the array names being mixed up. This has been fixed using
Ordered dictionaries.