ENH: Iterate over real particles.
Iterating over real particles makes parallel runs much more efficient as
the ghost and remote particles are not iterated over.  There are cases
where one may want to iterate over all particles for example when
computing the summation density or the TaitEOS, in these case those
equations should be put in a separate Group with the kwarg `real=False`
as done for the `dam_break.py` example.