ENH: No need to specify default args in equations.
Earlier, all predefined quantities needed to be supplied with default
values for the `CythonGenerator` to pick up the right types.  This was a
bit of a pain and is now removed.  The types are automatically
determined from the precomputed symbols themselves.  This is done by
passing a `known_types` argument to the `CythonGenerator`.