ENH: Adding configurable output arrays for particles
This change adds provides the user with the ability to set the arrays
that will be written to file. This should substantially reduce the
output file size.

Additionally, for parallel runs, the new default is to save out only
real particles. This can be changed by the
Solver.set_output_printing_level method