ENH: Add support for a `_get_helpers_` method.
An equation can define a `_get_helpers_` method that returns a sequence
of wrappable functions (pure Python), this is automatically wrapped to
high-performance Cython and added to the sph_eval code.  This minimizes
our need for the `_cython_code_` as the generator can be used to write
other backend code instead of Cython if needed.  The
`MonaghanKajtarBoundaryForce` now no longer uses `_cython_code_`.