ENH: Moving the periodicity functions from NNPS to the DomainManager
The periodicity routines should be called before the particle binning
process. Currently it is done within the NNPS.update method which is
not advisable since we may want to call the periodicity handling
routine as a post-stage callback.

I've extracted the periodicity functions into DomainManager and we
need to converge on a way to actually do this through the API.