Add interpolation support to the pysph_viewer.
Added a panel to the pysph_viewer that allows the user to nicely
interpolate the data onto a mesh and view that.  This should work in 2D
and 3D.  In 3D it can be slow due to the number of interpolations.
Modified the get_bounding_box to also stretch the domain by a small
amount if requested (which is the default).