Fix issue with missing points.
It seems mlab's vector_scatter has weird problems and some points are
not displayed when we set the u, v and w attributes, so we simply use
scalar_scatter for now which works without issue.  We will need to find
another way to see how to display vectors later on.