ENH: Support for iterating groups.
With this one can have a group iterate based on a convergence criterion.
The group should set the iterate attribute to True and each equation
should define a `converged` method that returns a positive value  to
indicate convergence.  The group's equations will be run multiple times
until convergence is reached.  See the `TestAccelerationEval1D` for an
example.