BUG: Correcting the iterative density update.
The running sum for the kernel derivative w.r.t. h was being
calculated without the respective array being initialized. This was
affecting the convergence properties of the iterative scheme.

Also, the omega or grad-h terms should be initialized in the
integrator. That is, at the beginning of a group, like the converged
attribute.