WIP: Fixes to IISPH implementation.
Some minor but important fixes.  Added a `get_particle_array_iisph` to
`base.utils`.  Reuse `db_geometry` code in examples to work for IISPH
examples.  Add a broken version of boundary support for IISPH.  A simple
example of two blocks of fluid interacting has been added for testing
the implementation.  Added support for adaptive timesteps in IISPH.  The
boundary support does not work and is WIP.