BUG: Correcting the velocity extrapolation for the no-slip BC
Calculation of the smoothed velocity close to the
boundary was incorrect. Volume for the particles is
calculated using both: fluid and wall particles. This
cannot be used for the normalisation while smoothing.
The kernel summation factor must be evaluated using
only the fluid particles.

This fixes the no-slip BC which was earlier not being
implemeneted correctly. Changes need to be made to
the other examples as well.