Split reduce into parallel and serial versions.
The trouble with a single reduce array is that each reduction is an
all-to-all communication.  We can make this more efficient by having a
`serial_reduce_array` be called and then call one single
`parallel_reduce_array` call to reduce an array of values across
processors.  This is what is recommended now.