Fix some issues with IISPH implementation.
This checkin changes the viscosity implementation to be a little more
standard as per Morris' formula.  The adaptive force calculation is also
fixed. The IISPH implementation currently has a lot of issues and is
erratic at best especially when considering solid boundaries.  These
changes are temporary and a proper investigation of the issues is
needed.  This will be done later, so use this at your own risk
currently.