BUG: Fix and improve damping of timesteps.
Previously, the damping performed resulted in very small timesteps that
were incorrect.  Here we change the implementation such that the
timesteps are damped (if required by specifying a non-zero n_damp), for
the first n_damp iterations.  Otherwise, no damping is performed.
Earlier, the first timestep was not damped and was not chosen via an
adaptive timestep, this has been fixed.  The solve method is also
cleaned up.