More general motion of inlet particles.
This allows the motion of the particles in the inlet to be in any
direction.  The only constraint is that the inlet region is
axis-aligned.  In the previous implementation the inlet particles always
had to stream along the positive x-axis.  This is no longer needed, the
particles could move via the face normal to the negative x-axis or any
other axis also.