BUG: Fix parallel adaptive timesteps.
In a parallel simulation if one of the processors does not have any
particles, then it has no timestep constraints and the implementation
was such that it would return the default timestep which is incorrect
(and often too low).  We instead return None if there are no constraints
and let the other processors determine the timestep.  There is no change
to the behavior in the serial case.