BUG: Inlet spacing and domain must be consistent.
The inlet must always be created such that along the stacked direction
`n * spacing` should equal the size of the domain.  Without this, one
would get particles on top of each other or other problems.  This
checkin adds a simple check for this case and raises an exception to the
user.  It also fixes/improves the trivial example to show a working
summation density.