Add a depends argument to `ExtModule`.
This allows us to automatically recompile modules if dependencies have
changed.  Updated the `SPHCompiler` so it uses `nnps` as a dependency.
Given the recent changes to `setup.py`, this will ensure that generated
modules are automatically recompiled when carray/particle_array or nnps
change.  This is a major pain point for users who end up with strange
errors and need to remove their `~/.pysph/source/` directory.