Implement box sort nnps efficiently.
The earlier `BoxSortNNPS` is now called `DictBoxSortNNPS`.  The new
`BoxSortNNPS` reuses a lot of the `LinkedListNNPS` class.  It simply
adds a mapping from cell index to an actual index to optimize the
storage of cells, the Linked List implementation would store all empty
cells.  The mapping is done using the C++ `std::map` which is very
performant and does not require the GIL allowing us to use `BoxSortNNPS`
with OpenMP as well.  Since nnps now requires C++, we make cython
generate C++ for all extensions for consistency.