Add option `--sort-gids` to sort neighbors.
This allows us to compare serial and parallel runs and check if
they are consistent.  With this, the parallel tests run with high
accuracy.  Also added tests for the sorting.  The sorting slows things
down by 25% or more.  The sorting is done by the NNPS and any example
can use this by simply passing `--sort-gids`.