Adding tox setup to test on py26, py27 and py34.
Also fixes some issues with the setup.py and sdist not working
optimally.  sdist was bundling the c++ extensions generated but these
may have been generated using a different setup (with or without OpenMP)
and this would result in strange errors for users.  We instead rely on
the build always generating fresh extension files. [reset minion]