Write an info file for convenience.
This info file can be used to post-process data.  Right now it contains
information on the output directory, output filename, arguments passed,
wall CPU time taken and if the run has completed or not..  Also add a
dummy method for post-processing.  This is important for it lets users
define the problem and also perform the analysis in the same script.
This analysis code can be reused easily.  Added a property called
`output_files` which returns all the output files.   Cleaned up the
application module.