Add complete examples into main sources.
This allows us to bundle the examples with the code, making it easier
for the users.  The examples use the new `Application` API and allow us
to make the example completely self-contained including the
post-processing.  This makes it very easy to extend the examples and
reuse the code.  There is also a convenient runner program that lets
users list and choose the example to run.  This can be run using:
`python -m pysph.examples.run`.