API: Do not make sources=None default.
This allows us to specify mandatory arguments for an equation.  Also
make many of the constructor arguments mandatory.  This is important as
a user may forget to pass an important parameter leading to subtle
errors.  This does change the API but hopefully the changes to user code
are very minimal.