Add option to initialize particles.
This takes an output file from another run and can initialize the run
with those particles.  This allows us to reproduce the results from the
TVF paper where a relaxed initialization is done by using the cartesian
grid results as input for the initial distribution of particles.  We
also use the QuinticSpline now as this produces much better results over
time than the CubicSpline.  This is also what the TVF paper uses.