Test that all examples run.
Added a test that simply takes one timestep for each available example.
This is a slow test as the compilation takes time.  This checkin also
adds a --max-steps cmdline argument to the application.  Fixed some
issues with the examples so all of them execute correctly.  Note that
this does not check that the results are correct, it simply tests if the
example runs.