Finally fix the issue with failing tests on win32.
There was an unnecessary include path specified in ext_module which was
equal to the empty string (when the tests are run from the root of the
source dir), this resulted in a "-I" with no directory specified.  This
makes the compiler ignore subsequent include path specifications leading
to strange errors.  Also remove the win32 setup.py env var additions
which are unnecessary.