WIP: First cut of SPH Schemes.
This adds a Scheme class that can be customized for the different SPH
schemes.  This will allow examples (and users) to simply choose and
configure a particular scheme which will generate the equations and
solver correctly.  This checkin adds a basic `WCSPHScheme` and has a few
examples working with this.  The scheme supports command line arguments
and a variety of useful options to test different schemes.  More schemes
should be implemented and more examples ported to use this.  The
application's command line help now lists all the default values for
convenience.