BUG: Viewing vmag with multiple arrays was broken.
If vmag was not already generated and the user is using play, the vmag
data is generated as soon as the first arrays particle array is updated
(at this point the other arrays aren't loaded) this would cause the file
to be dumped with a mix of old and new data.  This is now fixed and
should be more efficient.  Also added an option to change the play delay
and reduced the default to be faster.  Clicking on loop now will
automatically restart playing if it was already on.