Change example defaults to be similar to SPHysics
This changes the `dam_break_2d` and `dambreak_sphysics` examples so they
are similar to SPHysics examples (case1 and case5).  This allows for
easier comparisons.  This also fixes a bug with the 2d dam break where
beta was wrongly set to 2 which is incorrect.  This also increases tank
height to 4 as is done in SPHysics.