NEW: job management tools and improved automation.
The `pysph.tools.jobs` module allows a user to create a simple cluster
of worker machines on which arbitrary jobs can be submitted and
monitored.  This requires `psutil` to be installed. `execnet` is needed
for remote machines.  The `pysph.tools.automation` module now has its
own very simple automation management code to execute tasks using the
`jobs` module.  This eliminates the need for `luigi` or other tools.
The API is insipired by the luigi Task API.  This commit also adds tests
for the jobs and automation module.  The `test_automation` provides a
useful example demonstrating the use of the automation framework.