Add convenient automation related code.
This pulls out important functionality from a complex working prototype.
The `automate` function makes it almost trivial to automate a collection
of `Problem` instances.  Given a simulation directory, output directory
and a collection of problem classes it will automatically run all of
these.  Updated the tests and also made some of the tests a bit more
robust while waiting for a remote task to complete.