Fix test error for ext_module.
The test was creating the file after the mocking was done, this was
causing problems as in the time before the file was written another
process could have written.  Instead we now do the writing as a side
effect so the lock is still held by ExtModule and other processes cannot
write until that is done.