Fix bug with scheme chooser.
If one used multiple schemes, the defaults would get messed up.  For
example, with the EDAC scheme the default value of alpha was 0.0.
However, this clobbered the WCSPH default causing havoc.  We now set the
scheme defaults to None so we know if something is actually set by the
user.  This way the default values supplied in the code are honored but
still configurable via the command line.