Cleanup.
Remove some repeated code, add a few queue.finish statements.  Also pull
x, y, z to the CPU so the bounds are computed correctly for now.  There
are some bugs in the GPU NNPS code for multiple particle arrays that
needs to be fixed.  The rest of the opencl stuff really does work now
barring the NNPS bugs which shall be fixed later.