The use_double is now configured in one place.
So using pysph.base.config is the only place to configure, there are no
other settings, this prevents subtle errors like earlier where the NNPS
can be explicitly passed something that is inconsistent with everything
else.  This commit therefore removes the kwarg for use_double with the
GPU nnps classes.