API: change DT_ADAPT implementation.
The current implementation was a mess and unsuitable for parallel
execution via OpenCL etc.  We now instead use the standard mechanisms to
compute adaptive timesteps.  dt_cfl, dt_force, and dt_visc are arrays
used to set the factors per particle and the largest of these is found
in the integrator.  This allows us to eliminate some dirty code in the
Cython code as well and this approach works perfectly with OpenCL as
well.  All existing equations with DT_ADAPT have been updated.  This
does change the API so if you have your own implementation you will need
to change it -- it should be rather easy to do so.