Support for non-pairwise interactions.
Currently, by defining a `loop` method on an equation, users can do
pairwise interactions.  However, often one wants all the neighbors for a
variety of reasons.  One example is non-pairwise interactions but there
are many other uses for this.  This is now possible by defining a
`loop_all` method which can be passed `NBRS` and `N_NBRS` which pass the
neighboring particle indices and the number of such neighbors.  This can
be used to do pairwise interactions also.

I have also allowed a user to pass a kernel to an equation method by
passing the special name `KERNEL`.  This allows a user to compute the
kernel/gradient with `KERNEL.kernel/KERNEL.gradient` etc.  This is very
convenient in some cases and in particular when using the `loop_all`
method.

This currently only works on the CPU.