Support for loop_all on GPUs.
This allows us to use loop_all methods in equations and have them work
on the GPU, a simple test with a pairwise interaction is also tested.
This also adds support to pass the kernel to loop and loop_all methods
of equations.  It does NOT add support to pass the kernel to other
methods as that will mean I need to pass the SPH kernel to all the
OpenCL kernels we have.  While this can be done easily, I am only adding
the basic feature here which ought to be enough.