BUG: Ensure that order of equations is preserved.
Currently this was preserved for the loops but not for the
initialize/post_loop/reduce etc.  This is technically a problem as we
are supposed to guarantee the order.  In practice this does not matter
usually but some equations have dependencies where this is critical.  We
now ensure that everything is called strictly in the order in which it
is defined.  Thanks to Dileep Menon for spotting this issue.