Support for function annotations.
The code generator will look for the annotations and prioritize those
over default arguments.  This is used for detecting the types of the
arguments and the return value.  If annotations are not present,
defaults are looked at.  Also supports LOCAL_MEM/GLOBAL_MEM address
space declarations.  Works on Cython and OpenCL.