API: Integrators explicitly call update_domain.
In the past, the application automatically setup an `nnps.update_domain`
call as a `post_stage` callback.  The `update_domain` sets up any
periodic or mirror ghosts and also updates the cell size based on the
smoothing length.  This is not widely known and a source of some
confusion.  It is also not easy to configure in a scheme for example.
There are cases where there is no need to update the domain.  This API
change makes this an explicit call that users are supposed to add in
their integrators.  Specifically this should be done in the integrator's
`one_timestep` method.  All standard PySPH integrators have been
updated.