ENH: Add `Application.customize_output`.
This method can be used to configure the output visualization.  The
pysph mayavi viewer now looks for a mayavi_config.py file which can be
generated by overloading Application.customize_output and calling
Application._mayavi_config(...).  What this does is to setup the default
visualization options so users don't have to do this each time.  Have
customized three of the 3D examples to show how this is done.