Fix subtle issue with automatic syncing on gpu.
The issue is that originally before we call `converged`, we'd sync the
equation data from the GPU to the host.  However, `reduce` and
`converged` are pure Python calls on the GPU so if you set `self.conv =
1` or something it sets the Python attribute and not the GPU one.  So
when the converged call synced the value from the GPU it would clobber
the Python attribute.  We therefore no longer do the syncing and it is
up to the user who will call `self._pull` in their reduce or converged
method.  Note that `self._pull` is not available on the CPU which is a
bit of a pain currently but relatively easy to work around. Setting the
GPU attribute is not typical so this is quite acceptable.