Do not start multi-processing interface by default.
This is hardly used by anyone and most folks are not aware of this
feature.  It is now not started by default but is easy enough to enable
using:  ``--multiproc auto`` to do what was done earlier by default, of
course particular authkey/host/port may always be specified as before.