More OpenMP usage in equations and integrators.
Earlier we only used OpenMP for the Equation's `loop` and `loop_all`
methods.  This is not optimal as the number of cores increase.  With
12-20 core machines being routinely available it makes sense to use
OpenMP as much as possible and so now all the equation methods,
`initialize`, `initialize_pair`, `post_loop` in addition to the original
`loop/loop_all` are executed in parallel.  In addition, the integrator
steps are also executed in parallel.  This should give us much better
scale up with more cores.  Note that in the code, the loops do not set
nogil=True since they are already in a block where the gil is released.
This PR requires the latest version of compyle with
https://github.com/pypr/compyle/pull/37 merged.