WIP: Add CPU support for `Group.loop_count`.
Currently all equations are iterated for all the destination particles.
One can iterate either over all real particles or all real and ghost
particles.  By passing `loop_count=n`, only n particles in the
destimation `will be iterated over.  This is useful when you have things
like rigid bodies.  `loop_count` can be either an integer or a string.
This works on the CPU and the chunksize is not set in this case to allow
an automatic setting instead of the default 64 which may be
inappropriate.