Changed loop_count to start_idx and stop_idx.
This allows us to specify the starting and ending indices which is very
convenient.  For example, if you only want to iterate over the ghost
particles which are at the end, the `start_idx` can be used and the
`stop_idx` serves the purpose of what used to be `loop_count`.