Add CPU support for a conditional group.
Groups now support a condition kwarg which takes a callable that is
passed t, dt.  If it returns True the group is executed, else it is not.
This is very convenient for certain schemes.  Also cleaned up the mako
code so it doesn't generate a bunch of "if True" statements in the
generated pyx files.