ENH: Support setting load balancing weights.
These can be set per particle array and are used for parallel execution.
So for example, the solid particles take a lot less computation so
should not be given the same weight as fluid particles.  In this case,
simply add a constant called `lb_weight` to the solid particle arrays
and set the value to 0.1 for example. This will improve the load
balancing.  Each particle array can be given a different weight.
The WCSPH Scheme has been updated to set the weights.  The others
require some investigation.