fix(NNPSs): fmax() and fmin() conflict
Builds were failing on windows github actions runner due to some
conflict involving fmax() and fmin().

Seems like  fmax() and fmin() were defined soleley for windows as
they are a part of C99 standard but earlier MSVC not support C99
features fully. But now, MSVC supports it. So, it is no longer required.